Generate report/requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.1
babel==2.16.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
docxcompose==1.4.0
docxtpl==0.18.0
Eel==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
plum-dispatch==1.7.4
pycparser==2.22
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-docx==1.1.2
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==72.2.0
six==1.16.0
Spire.Doc==12.7.1
typing_extensions==4.12.2
urllib3==2.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.0.1
